style : point에서 payment로 파일명 변경
