+++ src/router/Routers.tsx
@@ -8,14 +8,12 @@ const SignUp = lazy(() => import("pages/SignUp/SignUp"));
 const Main = lazy(() => import("pages/Main/Main"));
 const SignIn = lazy(() => import("pages/SignIn/SignIn"));
 const My = lazy(() => import("pages/My/My"));
-const Payment = lazy(() => import("pages/Payment/Payment"));
 const Find = lazy(() => import("pages/Find/Find"));
 const UsedBook = lazy(() => import("pages/UsedBook/UsedBook"));
 const Oauth = lazy(() => import("pages/Oauth/Oauth"));
 const BookDetail = lazy(() => import("pages/BookDetail/BookDetail"));
 const BookReviewList = lazy(() => import("pages/BookReviewList/BookReviewList"));
 const UsedBookDetail = lazy(() => import("pages/UsedBookDetail/UsedBookDetail"));
-const Point = lazy(() => import("pages/Point/Point"));
 
 const Routers = () => {
   const { signIn, dispatch } = useSignIn();
@@ -29,14 +27,11 @@ const Routers = () => {
   return (
     <Switch>
       <Route path="/" exact component={Main} />
-      <PrivateRoute path="/point" exact component={Point} redirectPath="/signIn" isLoggedIn />
       <PrivateRoute path="/signUp" component={SignUp} redirectPath="/" isLoggedIn={isLoggedIn} />
       <PrivateRoute path="/signIn" component={SignIn} redirectPath="/" isLoggedIn={isLoggedIn} />
       <PrivateRoute path="/my" component={My} redirectPath="/signIn" isLoggedIn={isLoggedIn} />
-      <PrivateRoute path="/payment" component={Payment} redirectPath="/signIn" isLoggedIn={isLoggedIn} />
       <Route path="/logout" render={handleLogout} />
       <Route path="/find" component={Find} />
-      <Route path="/usedBook/:id" render={() => <div>중고도서 상세 페이지</div>} />
       <Route path="/usedBook/:id" component={UsedBookDetail} />
       <Route path="/usedBook" component={UsedBook} />
       <Route path="/oAuth/:name" component={Oauth} />
